+++ clinica/converters/adni_to_bids/modality_converters/_dwi.py
@@ -48,7 +48,8 @@ def convert_dwi(
     """
     from clinica.utils.stream import cprint
 
-    from .._utils import ADNIModalityConverter, paths_to_bids
+    from .._modality import ADNIModalityConverter
+    from .._utils import paths_to_bids
 
     cprint(
         f"Calculating paths of DWI images. Output will be stored in {conversion_dir}.",
