+++ clinica/converters/adni_to_bids/modality_converters/_fdg_pet.py
@@ -15,7 +15,7 @@ from clinica.converters.adni_to_bids.modality_converters._pet_utils import (
     ADNITracer,
 )
 
-from .._utils import ADNIModalityConverter, ADNIPETPreprocessingStep
+from .._modality import ADNIModalityConverter, ADNIPETPreprocessingStep
 
 
 def convert_fdg_pet(
