+++ clinica/converters/adni_to_bids/modality_converters/_fmap.py
@@ -51,7 +51,8 @@ def convert_fmap(
     """
     from clinica.utils.stream import cprint
 
-    from .._utils import ADNIModalityConverter, paths_to_bids
+    from .._modality import ADNIModalityConverter
+    from .._utils import paths_to_bids
 
     csv_dir = Path(csv_dir)
     source_dir = Path(source_dir)
