+++ clinica/converters/adni_to_bids/modality_converters/_fmri.py
@@ -54,7 +54,8 @@ def convert_fmri(
     """
     from clinica.utils.stream import cprint
 
-    from .._utils import ADNIModalityConverter, paths_to_bids
+    from .._modality import ADNIModalityConverter
+    from .._utils import paths_to_bids
 
     cprint(
         f"Calculating paths of {ADNIModalityConverter.FMRI} images. Output will be stored in {conversion_dir}.",
